chapter8 base skeleton

+++ chapter8/football/src/main/java/com/packt/football/domain/Album.java
@@ -1,5 +1,5 @@
 package com.packt.football.domain;
 
-public record Album (Integer id, String title) {
+public record Album (Long id, String title) {
 
 }
