+++ chapter8/football/src/main/java/com/packt/football/domain/Player.java
@@ -2,6 +2,6 @@ package com.packt.football.domain;
 
 import java.time.LocalDate;
 
-public record Player(Integer id, String name, Integer jerseyNumber, String position, LocalDate dateOfBirth) {
+public record Player(Long id, String name, Integer jerseyNumber, String position, LocalDate dateOfBirth) {
     
 }
